Moved controller objects to seperate class and renamed source to producer
